refactor(infra): 비속어, 배달 클라이언트 테스트 클래스 패키지 정리
